requirements.txt
Django
djangorestframework
South
dj-database-url
psycopg2
gunicorn
django-uuidfield
